Moved NestedForLoop from basic to design ruleset
