+++ src/main/resources/rulesets/basic.xml
@@ -23,7 +23,6 @@
     <rule class='org.codenarc.rule.basic.ConstantIfExpressionRule'/>
     <rule class='org.codenarc.rule.basic.DeadCodeRule'/>
     <rule class='org.codenarc.rule.basic.DoubleNegativeRule'/>
-    <rule class='org.codenarc.rule.basic.NestedForLoopRule'/>
     <rule class='org.codenarc.rule.basic.DuplicateCaseStatementRule'/>
     <rule class='org.codenarc.rule.basic.DuplicateMapKeyRule'/>
     <rule class='org.codenarc.rule.basic.DuplicateSetValueRule'/>
